pyproject.toml
[project]
name = "discordbot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "discord-py>=2.6.4",
    "flask>=3.1.2",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
]
